third_party/WebKit/LayoutTests/fast/imagecapture/MediaStreamTrack-applyConstraints.html

<!DOCTYPE html>
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<body>
<canvas id='canvas' width=10 height=10/>
</body>
<script>

// This test verifies that MediaStreamTrack.applyConstraints() exists and that,
// when called with no parameters, returns a Promise that is resolved. This
// might not make sense: https://github.com/w3c/mediacapture-main/issues/438 .
// Other tests go deeper.
promise_test(function(t) {
  var canvas = document.getElementById('canvas');
  var context = canvas.getContext("2d");
  context.fillStyle = "red";
  context.fillRect(0, 0, 10, 10);

  var stream = canvas.captureStream();
  assert_equals(stream.getAudioTracks().length, 0);
  assert_equals(stream.getVideoTracks().length, 1);

  var videoTrack = stream.getVideoTracks()[0];
  assert_equals(typeof videoTrack.applyConstraints, 'function');

  return videoTrack.applyConstraints();
}, 'MediaStreamTrack.applyConstraints()');

</script>
